requirements.txt
# --- Core Backend ---
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.0.1

# --- Clinical AI & NLP ---
google-generativeai==0.8.4
spacy==3.8.4
pydantic[email]==2.12.5

# --- Database & Security ---
motor==3.7.0
pymongo==4.11.1
pyjwt==2.10.1
bcrypt==4.2.1

# --- Networking ---
httpx==0.28.1